batch: Move presence constraint realization
Batch merge currently owns chunk emission, candidate enumeration, discard
replay, and the routines that realize claimed source lines against a
working tree.

The project does not provide a separate owner for presence realization, so
callers that only need claimed-line replay must depend on merge workflow
internals.

This commit addresses that by moving presence realization into
batch.presence_constraints, moving shared working-range realization into
batch.realized_mapping, and pointing source advancement plus storage at the
presence owner.

Subsequent commits will update tests to assert the presence and realized
mapping boundaries.
